Add vsts control functions
