--- MoleDeploy.VstsControl/Properties/PublishProfiles/mtcden-sandbox-demo-whack-a-mole-vsts-func - Web Deploy.pubxml
@@ -1,25 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<!--
-This file is used by the publish/package process of your Web project.You can customize the behavior of this process
-by editing this MSBuild file. In order to learn more about this please visit https://go.microsoft.com/fwlink/?LinkID=208121. 
--->
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <WebPublishMethod>MSDeploy</WebPublishMethod>
-    <PublishProvider>AzureWebSite</PublishProvider>
-    <LastUsedBuildConfiguration>Release</LastUsedBuildConfiguration>
-    <LastUsedPlatform>Any CPU</LastUsedPlatform>
-    <SiteUrlToLaunchAfterPublish>http://mtcden-sandbox-demo-whack-a-mole-vsts-func.azurewebsites.net</SiteUrlToLaunchAfterPublish>
-    <LaunchSiteAfterPublish>False</LaunchSiteAfterPublish>
-    <ExcludeApp_Data>False</ExcludeApp_Data>
-    <MSDeployServiceURL>mtcden-sandbox-demo-whack-a-mole-vsts-func.scm.azurewebsites.net:443</MSDeployServiceURL>
-    <ResourceId>/subscriptions/b0e04a4a-a321-4b66-b8fd-13715262ba3c/resourceGroups/demo-whack-a-mole-sandbox-rg/providers/Microsoft.Web/sites/mtcden-sandbox-demo-whack-a-mole-vsts-func</ResourceId>
-    <DeployIisAppPath>mtcden-sandbox-demo-whack-a-mole-vsts-func</DeployIisAppPath>
-    <SkipExtraFilesOnServer>True</SkipExtraFilesOnServer>
-    <MSDeployPublishMethod>WMSVC</MSDeployPublishMethod>
-    <EnableMSDeployBackup>True</EnableMSDeployBackup>
-    <UserName>$mtcden-sandbox-demo-whack-a-mole-vsts-func</UserName>
-    <_SavePWD>True</_SavePWD>
-    <EnableMsDeployAppOffline>False</EnableMsDeployAppOffline>
-  </PropertyGroup>
-</Project>
