+++ MoleDeploy.VstsCore/.gitignore
