+++ MoleDeploy.VstsCore/Functions/ReportBuildState.cs
