+++ MoleDeploy.VstsCore/MoleDeploy.VstsCore.csproj
@@ -4,13 +4,12 @@
     <AzureFunctionsVersion>v2</AzureFunctionsVersion>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Sdk.Functions" Version="1.0.14" />
-    <!--<PackageReference Include="Microsoft.VisualStudio.Services.Client" Version="16.135.0-preview" />
-    <PackageReference Include="Microsoft.VisualStudio.Services.InteractiveClient" Version="16.135.0-preview" />
-    <PackageReference Include="Microsoft.TeamFoundationServer.Client" Version="16.135.0-preview" />
-    <PackageReference Include="Microsoft.VisualStudio.Services.Release.Client" Version="16.135.0-preview" />-->
-    <PackageReference Include="Newtonsoft.Json" Version="11.0.2" />
-    <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="5.2.4" />
+    <PackageReference Include="Microsoft.NET.Sdk.Functions" Version="1.0.13" />
+    <PackageReference Include="Microsoft.VisualStudio.Services.Client" Version="15.133.0-preview" />
+    <PackageReference Include="Microsoft.VisualStudio.Services.InteractiveClient" Version="15.133.0-preview" />
+    <PackageReference Include="Microsoft.TeamFoundationServer.Client" Version="15.133.0-preview" />
+    <PackageReference Include="Microsoft.VisualStudio.Services.Release.Client" Version="15.133.0-preview" />
+    <PackageReference Include="Newtonsoft.Json" Version="10.0.3" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\MoleDeploy.Contracts\MoleDeploy.Contracts.csproj" />
