+++ MoleDeploy.VstsCore/VstsClient.cs
