+++ MoleDeploy.VstsCore/host.json
