+++ WhackAMole.sln
@@ -22,6 +22,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MoleDeploy.SignalR", "MoleD
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MoleDeploy.Vsts", "MoleDeploy.VstsControl\MoleDeploy.Vsts.csproj", "{BE6213F2-22FD-4446-B6ED-242B3B32E779}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MoleDeploy.VstsCore", "MoleDeploy.VstsCore\MoleDeploy.VstsCore.csproj", "{21D5A425-11D0-4EEB-8CE4-9135B50AA099}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -158,6 +160,22 @@ Global
 		{BE6213F2-22FD-4446-B6ED-242B3B32E779}.Release|x64.Build.0 = Release|Any CPU
 		{BE6213F2-22FD-4446-B6ED-242B3B32E779}.Release|x86.ActiveCfg = Release|Any CPU
 		{BE6213F2-22FD-4446-B6ED-242B3B32E779}.Release|x86.Build.0 = Release|Any CPU
+		{21D5A425-11D0-4EEB-8CE4-9135B50AA099}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{21D5A425-11D0-4EEB-8CE4-9135B50AA099}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{21D5A425-11D0-4EEB-8CE4-9135B50AA099}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{21D5A425-11D0-4EEB-8CE4-9135B50AA099}.Debug|ARM.Build.0 = Debug|Any CPU
+		{21D5A425-11D0-4EEB-8CE4-9135B50AA099}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{21D5A425-11D0-4EEB-8CE4-9135B50AA099}.Debug|x64.Build.0 = Debug|Any CPU
+		{21D5A425-11D0-4EEB-8CE4-9135B50AA099}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{21D5A425-11D0-4EEB-8CE4-9135B50AA099}.Debug|x86.Build.0 = Debug|Any CPU
+		{21D5A425-11D0-4EEB-8CE4-9135B50AA099}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{21D5A425-11D0-4EEB-8CE4-9135B50AA099}.Release|Any CPU.Build.0 = Release|Any CPU
+		{21D5A425-11D0-4EEB-8CE4-9135B50AA099}.Release|ARM.ActiveCfg = Release|Any CPU
+		{21D5A425-11D0-4EEB-8CE4-9135B50AA099}.Release|ARM.Build.0 = Release|Any CPU
+		{21D5A425-11D0-4EEB-8CE4-9135B50AA099}.Release|x64.ActiveCfg = Release|Any CPU
+		{21D5A425-11D0-4EEB-8CE4-9135B50AA099}.Release|x64.Build.0 = Release|Any CPU
+		{21D5A425-11D0-4EEB-8CE4-9135B50AA099}.Release|x86.ActiveCfg = Release|Any CPU
+		{21D5A425-11D0-4EEB-8CE4-9135B50AA099}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
